+++ examples/block-handler-vaults/manifest.ts
@@ -1,4 +1,4 @@
-import { Manifest } from 'https://deno.land/x/robo_arkiver@v0.4.15/mod.ts'
+import { Manifest } from 'https://deno.land/x/robo_arkiver@v0.5.0/mod.ts'
 import { VaultSnapshot } from './entities/vault.ts'
 import { snapshotVault } from './handlers/vault.ts'
 
@@ -6,7 +6,7 @@ const manifest = new Manifest('yearn-vaults')
 
 manifest
   .addEntity(VaultSnapshot)
-  .chain('mainnet')
+  .addChain('mainnet')
   .addBlockHandler({
     blockInterval: 1000,
     startBlockHeight: 12790000n,