+++ examples/simple/manifest.ts
@@ -7,8 +7,8 @@ const manifest = new Manifest('simple')
 
 manifest
   .addEntity(Balance)
-  .chain('mainnet', { blockRange: 100n })
-  .contract('ERC20', erc20)
+  .addChain('mainnet', { blockRange: 100n })
+  .addContract('ERC20', erc20)
   .addSources({ '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2': 16987011n })
   .addEventHandlers({ 'Transfer': transferHandler })